Time: 6 ms (91.07%), Space: 54.9 MB (62.44%) - LeetHub

+++ 1975-maximum-matrix-sum/1975-maximum-matrix-sum.java
@@ -16,9 +16,9 @@ class Solution {
             }
             }
             if(count%2==0){
-                return (long)sum;
+                return sum;
             }else{
-                return (long)sum-2*minVal;
+                return sum-2*minVal;
             }
         
     }
